refactor: move main components to screen folder
As some components are not global and are the main components to build our screen structure, i
created a screen folder containing those elements while keeping only shareable components inside
components folder.
