--- src/components/Connection/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-import { Container } from './styles'
-
-const Connection: React.FC = ({ children }) => {
-  return <Container>{children}</Container>
-}
-
-export default Connection
