--- src/components/Connection/styles.ts
@@ -1,6 +1,0 @@
-import styled from 'styled-components'
-
-export const Container = styled.div`
-  flex: 1;
-  display: flex;
-`
