+++ src/components/ToastContainer/index.tsx
@@ -3,8 +3,8 @@ import { createPortal } from 'react-dom'
 import { useTransition } from 'react-spring'
 
 import { ToastMessage } from '../../context/toast'
-import Toast from '../Toast'
 import { Container } from './styles'
+import Toast from './Toast'
 
 interface ToastContainerProps {
   toasts: ToastMessage[]
