+++ src/screen/ConnectionsList/NewConnectionModal/styles.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-import Button from '../Button'
+import Button from '../../../components/Button'
 
 export const ActionsContainer = styled.div`
   display: flex;
